Task 1/Deliverables/requirements.txt
docx_mailmerge==0.5.0
PyYAML==6.0.1
